feature: like and unlike books

+++ src/main/java/com/example/demo/common/UID.java
@@ -6,15 +6,15 @@ import lombok.Getter;
 @Getter
 public class UID {
     @JsonIgnore
-    private Long localId;
+    private Integer localId;
 
     @JsonIgnore
-    private Long objectType;
+    private Integer objectType;
 
     @JsonIgnore
-    private int shardID;
+    private Integer shardID;
 
-    public UID(Long localId, Long objectType, int shardID) {
+    public UID(Integer localId, Integer objectType, Integer shardID) {
         this.localId = localId;
         this.objectType = objectType;
         this.shardID = shardID;
@@ -22,14 +22,14 @@ public class UID {
 
     @Override
     public String toString() {
-        Long val = localId << 28 | objectType << 18 | Long.valueOf(shardID) << 0;
+        Long val = Long.valueOf(localId) << 28 | Long.valueOf(objectType) << 18 | Long.valueOf(shardID) << 0;
         return Long.toString(val);
     }
 
     public static UID DecomposeUID(String s) {
         Long uid = Long.valueOf(s);
 
-        UID u = new UID(uid >> 28, uid >> 18 & 0x3ff, (int)(uid >> 0 & 0x3FFFF));
+        UID u = new UID((int)(uid >> 28), (int)(uid >> 18 & 0x3ff), (int)(uid >> 0 & 0x3FFFF));
         return u;
     }
 }
